updating edit services
