+++ controllers/index.js
@@ -2,9 +2,11 @@ const router = require('express').Router();
 
 const apiRoutes = require('./api');
 const homeRoutes = require('./home-routes.js');
+const adminRoutes = require('./admin-routes.js');
 
 router.use('/api', apiRoutes);
 router.use('/', homeRoutes);
+router.use('/admin', adminRoutes);
 
 router.use((req, res) => {
   res.status(404).end();
