+++ views/edit-services.handlebars
@@ -1,5 +1,53 @@
 <div class="row">
     <div class="col-12 mt-3">
+        <h2 class="page-title">Add New Service</h2>
+        <form class="service-form">
+            <div>
+                <div class="row">
+
+               
+                <div class="col-6">
+                <div>
+                    <label for="category">category:</label>
+                    <select type="text" id="category-create" class="form-control">
+                        <option id="" value="">Choose a Category</option>
+                        <option id="braids" value="braids">Braids</option>
+                        <option id="cuts" value="cuts">Cuts</option>
+                        <option id="color" value="color">Color</option>
+                        <option id="perms" value="perms">Perms</option>
+                    </select>
+                </div>
+                <div>
+                    <label for="style">style name:</label>
+                    <input type="text" id="style-create" class="form-control"/>
+                </div>
+                <div>
+                    <label for="time_alloted">time needed for style:</label>
+                    <input type="time" id="time-create" class="form-control"/>
+                </div>
+                
+                </div>
+                <div class="col-6">
+                <div>
+                    <label for="description">description:</label>
+                    <textarea class="form-control" name="description-create"></textarea>
+                </div>
+                
+                <div>
+                    <label for="price">price of service:</label>
+                    <input type="decimal" id="price-create" class="form-control"/>
+                </div>
+                </div>
+                 </div>
+                <br>
+                <div>
+                   <button class="btn float-right" type="submit">Create New Service</button>
+                </div>
+            </div>
+        </form>
+        <br>
+        <hr>
+
         <h2 class="page-title">Choose Service to Edit</h2>         
         <div class="card-deck">
          
@@ -22,7 +70,7 @@
                         This style takes: {{time_alloted}} <br>
                         ${{price}}
                     </div>
-                    <div class="col-2 d-flex"><button class="btn justify-content-center align-self-center">Edit</div>
+                    <div class="col-2 d-flex"><button class="btn justify-content-center align-self-center"><a href="/admin/edit/{{id}}">Edit</a></div>
                 </div>
                  {{/each}}
             </div> <!--end card -->
