+++ views/layouts/main.handlebars
@@ -9,7 +9,7 @@
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/css/bootstrap.min.css">
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.14.0/css/all.min.css" integrity="sha512-1PKOgIY59xJ8Co8+NE6FZ+LOAZKjy+KY8iq0G4B3CyeY6wYHN3yt9PW0XpSriVlkMXe40PTKnXrLnZ9+fkDaog==" crossorigin="anonymous" />
 
-        <link rel="stylesheet" href="stylesheets/style.css" />
+        <link rel="stylesheet" href="/stylesheets/style.css" />
      
   </head>
   <body>
@@ -22,11 +22,11 @@
                 <a class="nav-link" href="/">Home</a>
               </li>
               <li class="nav-item">
-                <a class="nav-link" href="#">Dashboard</a>
+                <a class="nav-link" href="/dashboard">Dashboard</a>
               </li>
               <!--if the stylist is logged in, this admin button will show else it wont-->
               <li class="nav-item">  
-                <a class="nav-link" href="#">Admin</a>
+                <a class="nav-link" href="/admin">Admin</a>
               </li>
               <li class="nav-item">
                 <a class="nav-link" href="/login">Login</a>
